+++ fiddles/nativeScript/fiddle-0009-SqliteDb/puzzle/platforms/ios/puzzle/app/app.js
@@ -1,5 +1,0 @@
-var application = require("application");
-
-application.cssFile = "./app.css";
-application.title = "15 Puzzle";
-application.start({ moduleName: "view/main-page" });