+++ fiddles/nativeScript/fiddle-0009-SqliteDb/puzzle/platforms/ios/puzzle/app/shared/levels.js
@@ -1,6 +1,0 @@
-
-module.exports = {
-  1: "view/level-one/level-one",
-  2: "view/level-two/level-two",
-  3: "view/level-three/level-three"
-};