+++ fiddles/nativeScript/fiddle-0010-Groceries/groceries/app/app.component.js
@@ -1,19 +1,43 @@
 "use strict";
 var core_1 = require("@angular/core");
+var user_1 = require("./shared/user/user");
+var user_service_1 = require("./shared/user/user.service");
 var AppComponent = (function () {
-    function AppComponent() {
-        this.email = "nativescriptrocks@telerik.com";
+    function AppComponent(userService) {
+        this.userService = userService;
+        this.isLoggingIn = true;
+        this.user = new user_1.User();
     }
     AppComponent.prototype.submit = function () {
-        alert("You’re using: " + this.email);
+        if (this.isLoggingIn) {
+            this.login();
+        }
+        else {
+            this.signUp();
+        }
+    };
+    AppComponent.prototype.login = function () {
+        // TODO: Define
+    };
+    AppComponent.prototype.signUp = function () {
+        var _this = this;
+        this.userService.register(this.user)
+            .subscribe(function () {
+            alert("Your account was successfully created.");
+            _this.toggleDisplay();
+        }, function () { return alert("Unfortunately we were unable to create your account."); });
+    };
+    AppComponent.prototype.toggleDisplay = function () {
+        this.isLoggingIn = !this.isLoggingIn;
     };
     AppComponent = __decorate([
         core_1.Component({
             selector: "my-app",
-            templateUrl: "app.component.html",
-            styleUrls: ["pages/login/login-common.css", "pages/login/login.css"]
+            providers: [user_service_1.UserService],
+            templateUrl: "pages/login/login.html",
+            styleUrls: ["pages/login/login-common.css", "pages/login/login.css"],
         }), 
-        __metadata('design:paramtypes', [])
+        __metadata('design:paramtypes', [user_service_1.UserService])
     ], AppComponent);
     return AppComponent;
 }());