+++ fiddles/nativeScript/fiddle-0010-Groceries/groceries/app/app.component.ts
@@ -1,13 +1,7 @@
-import {Component} from "@angular/core";
+import { Component } from "@angular/core";
 
 @Component({
-    selector: "my-app",
-    templateUrl: "app.component.html",
-    styleUrls: ["pages/login/login-common.css", "pages/login/login.css"]
+    selector: "main",
+    template: "app.component.html"
 })
-export class AppComponent {
-    email = "nativescriptrocks@telerik.com";
-    submit() {
-        alert("You’re using: " + this.email);
-    }
-}
+export class AppComponent {}