+++ fiddles/nativeScript/fiddle-0010-Groceries/groceries/app/app.module.js
@@ -1,0 +1,25 @@
+"use strict";
+var core_1 = require("@angular/core");
+var forms_1 = require("nativescript-angular/forms");
+var http_1 = require("nativescript-angular/http");
+var platform_1 = require("nativescript-angular/platform");
+var app_component_1 = require("./app.component");
+var AppModule = (function () {
+    function AppModule() {
+    }
+    AppModule = __decorate([
+        core_1.NgModule({
+            imports: [
+                platform_1.NativeScriptModule,
+                forms_1.NativeScriptFormsModule,
+                http_1.NativeScriptHttpModule
+            ],
+            declarations: [app_component_1.AppComponent],
+            bootstrap: [app_component_1.AppComponent]
+        }), 
+        __metadata('design:paramtypes', [])
+    ], AppModule);
+    return AppModule;
+}());
+exports.AppModule = AppModule;
+//# sourceMappingURL=app.module.js.map