+++ fiddles/nativeScript/fiddle-0010-Groceries/groceries/app/main.js
@@ -1,20 +1,6 @@
 "use strict";
 // this import should be first in order to load some required settings (like globals and reflect-metadata)
 var platform_1 = require("nativescript-angular/platform");
-var core_1 = require("@angular/core");
-var app_component_1 = require("./app.component");
-var AppComponentModule = (function () {
-    function AppComponentModule() {
-    }
-    AppComponentModule = __decorate([
-        core_1.NgModule({
-            declarations: [app_component_1.AppComponent],
-            bootstrap: [app_component_1.AppComponent],
-            imports: [platform_1.NativeScriptModule],
-        }), 
-        __metadata('design:paramtypes', [])
-    ], AppComponentModule);
-    return AppComponentModule;
-}());
-platform_1.platformNativeScriptDynamic().bootstrapModule(AppComponentModule);
+var app_module_1 = require("./app.module");
+platform_1.platformNativeScriptDynamic().bootstrapModule(app_module_1.AppModule);
 //# sourceMappingURL=main.js.map