+++ fiddles/nativeScript/fiddle-0010-Groceries/groceries/app/main.ts
@@ -1,13 +1,7 @@
 // this import should be first in order to load some required settings (like globals and reflect-metadata)
 import { platformNativeScriptDynamic, NativeScriptModule } from "nativescript-angular/platform";
+import { AppModule } from "./app.module";
 import { NgModule } from "@angular/core";
-import { AppComponent } from "./app.component";
 
-@NgModule({
-    declarations: [AppComponent],
-    bootstrap: [AppComponent],
-    imports: [NativeScriptModule],
-})
-class AppComponentModule {}
 
-platformNativeScriptDynamic().bootstrapModule(AppComponentModule);
+platformNativeScriptDynamic().bootstrapModule(AppModule);