+++ fiddles/nativeScript/fiddle-0010-Groceries/groceries/app/pages/list/list.component.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+
+@Component({
+    selector: "list",
+    templateUrl: "pages/list/list.html",
+    styleUrls: ["pages/list/list-common.css", "pages/list/list.css"]
+})
+export class ListComponent {}